+++ src/QueryCat.Cli/Commands/PluginRemoveCommand.cs
@@ -22,7 +22,7 @@ internal class PluginRemoveCommand : BaseCommand
             var plugin = parseResult.GetRequiredValue(pluginArgument);
 
             applicationOptions.InitializeLogger();
-            using var root = await applicationOptions.CreateApplicationRootAsync();
+            await using var root = await applicationOptions.CreateApplicationRootAsync();
             await root.PluginsManager.RemoveAsync(plugin, cancellationToken);
         });
     }