+++ src/QueryCat.Cli/Commands/ServeCommand.cs
@@ -60,7 +60,7 @@ internal class ServeCommand : BaseCommand
             var allowedIPsSlots = parseResult.GetValue(allowedIPsSlotsOption);
 
             applicationOptions.InitializeLogger();
-            using var root = await applicationOptions.CreateApplicationRootAsync();
+            await using var root = await applicationOptions.CreateApplicationRootAsync();
             root.Thread.Options.AddRowNumberColumn = true;
             root.Thread.Options.SafeMode = safeMode;
             var webServer = new WebServer(root.Thread, new WebServerOptions